# docs/openapi/api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "后台管理接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/models": {
      "get": {
        "summary": "获取模型列表",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "系统"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/models/": {
      "get": {
        "summary": "获取所有模型元数据",
        "deprecated": false,
        "description": "👨‍💼 需要管理员权限（Admin）",
        "tags": [
          "模型管理"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "post": {
        "summary": "创建模型元数据",
        "deprecated": false,
        "description": "👨‍💼 需要管理员权限（Admin）",
        "tags": [
          "模型管理"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "put": {
        "summary": "更新模型元数据",
        "deprecated": false,
        "description": "👨‍💼 需要管理员权限（Admin）",
        "tags": [
          "模型管理"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "SessionAuth1": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session认证，通过登录接口获取"
      },
      "AccessToken1": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Access Token认证，格式: Bearer {access_token}，通过 /api/user/token 接口生成"
      },
      "NewApiUser1": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "用户ID请求头，必须与当前登录用户ID匹配，使用Session或AccessToken认证时必须提供"
      },
      "Combination222": {
        "group": [
          {
            "id": 573666
          },
          {
            "id": 573668
          }
        ],
        "type": "combination"
      },
      "Combination1122": {
        "group": [
          {
            "id": 573667
          },
          {
            "id": 573668
          }
        ],
        "type": "combination"
      },
...
  },
  "servers": [],
  "security": [
    {
      "Combination343": []
    },
    {
      "Combination1243": []
    }
  ]
}